Module sqlx_migrator::cli

Available on crate feature cli only.

Module for creating and running cli with help of migrator

Structs

Apply

CLI struct for apply subcommand

Revert

CLI struct for revert subcommand

Enums

SubCommand

Subcommand for sqlx migrator cli

Functions

run

Run full cli by parsing args with help of migrator. If you only need to add subcommand to your app than use SubCommand enum handle_subcommand function