Trait sqlx_migrator::migration::MigrationTrait

pub trait MigrationTrait<DB>: Send + Sync {
    fn app(&self) -> &str;
    fn name(&self) -> &str;

    fn parents(&self) -> Vec<Box<dyn MigrationTrait<DB>>>  { ... }
    fn operations(&self) -> Vec<Box<dyn OperationTrait<DB>>>  { ... }
    fn replaces(&self) -> Vec<Box<dyn MigrationTrait<DB>>>  { ... }
    fn run_before(&self) -> Vec<Box<dyn MigrationTrait<DB>>>  { ... }
    fn is_atomic(&self) -> bool { ... }
}

Trait for migration

Required Methods

fn app(&self) -> &str

Migration app name. Can be name of folder or library where migration is located

fn name(&self) -> &str

Migration name. Can be file name without extension

Provided Methods

fn parents(&self) -> Vec<Box<dyn MigrationTrait<DB>>>

Parents of migration (migrations that should be applied before this migration)

fn operations(&self) -> Vec<Box<dyn OperationTrait<DB>>>

Operation performed for migration (create, drop, etc.)

fn replaces(&self) -> Vec<Box<dyn MigrationTrait<DB>>>

Replace certain migrations. If any one of listed migration is applied than migration will not be applied else migration will apply instead of applying those migration.

fn run_before(&self) -> Vec<Box<dyn MigrationTrait<DB>>>

Run before certain migration. This can be helpful in condition where other library migration need to be applied after this migration

fn is_atomic(&self) -> bool

Whether migration is atomic or not. By default it is true