Module sqlx_migrator::migrator

Migrator module

It contains common enum and trait for implementing migrator for sqlx supported database

It also provides its own struct Migrator which supports Postgres, Sqlite and Mysql Database only

Example

Create own custom Migrator which only supports postgres and uses own unique table name instead of default table name

use std::collections::HashSet;

use sqlx::{Pool, Postgres};
use sqlx_migrator::error::Error;
use sqlx_migrator::migration::{AppliedMigrationSqlRow, Migration};
use sqlx_migrator::migrator::{DatabaseOperation, Info, Migrate};

pub struct CustomMigrator {
    migrations: HashSet<Box<dyn Migration<Postgres>>>,
    pool: Pool<Postgres>,
}

impl Info<Postgres> for CustomMigrator {
    fn migrations(&self) -> &HashSet<Box<dyn Migration<Postgres>>> {
        &self.migrations
    }

    fn migrations_mut(&mut self) -> &mut HashSet<Box<dyn Migration<Postgres>>> {
        &mut self.migrations
    }

    fn pool(&self) -> &Pool<Postgres> {
        &self.pool
    }
}
#[must_use]
pub(crate) fn create_migrator_table_query() -> &'static str {
    "CREATE TABLE IF NOT EXISTS _custom_migrator_table_name (
       id INT PRIMARY KEY NOT NULL GENERATED ALWAYS AS IDENTITY,
       app TEXT NOT NULL,
       name TEXT NOT NULL,
       applied_time TIMESTAMPTZ NOT NULL DEFAULT now(),
       UNIQUE (app, name)
   )"
}

#[must_use]
pub(crate) fn drop_table_query() -> &'static str {
    "DROP TABLE IF EXISTS _custom_migrator_table_name"
}

#[must_use]
pub(crate) fn fetch_row_query() -> &'static str {
    "SELECT id, app, name, applied_time FROM _custom_migrator_table_name"
}

#[must_use]
pub(crate) fn add_migration_query() -> &'static str {
    "INSERT INTO _custom_migrator_table_name(app, name) VALUES ($1, $2)"
}

#[must_use]
pub(crate) fn delete_migration_query() -> &'static str {
    "DELETE FROM _custom_migrator_table_name WHERE app = $1 AND name = $2"
}

pub(crate) async fn lock_id(pool: &Pool<Postgres>) -> Result<i64, Error> {
    let (database,): (String,) = sqlx::query_as("SELECT CURRENT_DATABASE()")
        .fetch_one(pool)
        .await?;
    Ok(i64::from(crc32fast::hash(database.as_bytes())))
}

pub(crate) fn lock_database_query() -> &'static str {
    "SELECT pg_advisory_lock($1)"
}

pub(crate) fn unlock_database_query() -> &'static str {
    "SELECT pg_advisory_unlock($1)"
}

#[async_trait::async_trait]
impl DatabaseOperation<Postgres> for CustomMigrator {
    async fn ensure_migration_table_exists(&self) -> Result<(), Error> {
        sqlx::query(create_migrator_table_query())
            .execute(&self.pool)
            .await?;
        Ok(())
    }

    async fn drop_migration_table_if_exists(&self) -> Result<(), Error> {
        sqlx::query(drop_table_query()).execute(&self.pool).await?;
        Ok(())
    }

    async fn add_migration_to_db_table(
        &self,
        migration: &Box<dyn Migration<Postgres>>,
        connection: &mut <Postgres as sqlx::Database>::Connection,
    ) -> Result<(), Error> {
        sqlx::query(add_migration_query())
            .bind(migration.app())
            .bind(migration.name())
            .execute(connection)
            .await?;
        Ok(())
    }

    async fn delete_migration_from_db_table(
        &self,
        migration: &Box<dyn Migration<Postgres>>,
        connection: &mut <Postgres as sqlx::Database>::Connection,
    ) -> Result<(), Error> {
        sqlx::query(delete_migration_query())
            .bind(migration.app())
            .bind(migration.name())
            .execute(connection)
            .await?;
        Ok(())
    }

    async fn fetch_applied_migration_from_db(
        &self,
    ) -> Result<Vec<AppliedMigrationSqlRow>, Error> {
        let rows = sqlx::query_as(fetch_row_query())
            .fetch_all(&self.pool)
            .await?;
        Ok(rows)
    }

    async fn lock(
        &self,
        connection: &mut <Postgres as sqlx::Database>::Connection,
    ) -> Result<(), Error> {
        let lock_id = lock_id(&self.pool).await?;
        sqlx::query(lock_database_query())
            .bind(lock_id)
            .execute(connection)
            .await?;
        Ok(())
    }

    async fn unlock(
        &self,
        connection: &mut <Postgres as sqlx::Database>::Connection,
    ) -> Result<(), Error> {
        let lock_id = lock_id(&self.pool).await?;
        sqlx::query(unlock_database_query())
            .bind(lock_id)
            .execute(connection)
            .await?;
        Ok(())
    }
}

impl Migrate<Postgres> for CustomMigrator {}

Structs

Migrator

Migrator struct which store migrations graph and information related to different library supported migrations

Plan

Struct which determine type of plan to use

Enums

PlanType

Type of plan which needs to be generate

Traits

DatabaseOperation

Trait which is implemented for database for performing database related actions on database. Usually this trait is implemented for database to support certain database along with info trait

Info

Info trait which implements some of database agnostic methods to return data. Only required methods needs to be implemented if you want to create your own migrator struct. This trait implements methods which doesn’t depends on Database Operation trait methods

Migrate

Migrate trait which migrate a database according to requirements. This trait implements all methods which depends on DatabaseOperation trait. This trait is not required to implement any method since all have provided implementation and good to go for database agnostic case