Struct sqlx_migrator::migrator::Migrator

pub struct Migrator<DB> { /* private fields */ }

Migrator struct which store migrations graph and information related to different library supported migrations

Implementations

impl<DB> Migrator<DB>

pub fn with_prefix(self, prefix: impl Into<String>) -> Result<Self, Error>

Use prefix for migrator table name only ascii alpha numeric and underscore characters are supported for table name. prefix will set table name as _{prefix}{default_table_name} where default table name is _sqlx_migrator_migrations

Errors

When passed prefix is not ascii alpha numeric or underscore character

pub fn table_name(&self) -> &str

Get name of table which is used for storing migrations related information in database