Trait sqlx_migrator::migrator::DatabaseOperation

pub trait DatabaseOperation<DB>
where
    DB: Database,
{
    // Required methods
    fn ensure_migration_table_exists<'life0, 'life1, 'async_trait>(
        &'life0 self,
        connection: &'life1 mut <DB as Database>::Connection
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn drop_migration_table_if_exists<'life0, 'life1, 'async_trait>(
        &'life0 self,
        connection: &'life1 mut <DB as Database>::Connection
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn add_migration_to_db_table<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        migration: &'life1 Box<dyn Migration<DB>>,
        connection: &'life2 mut <DB as Database>::Connection
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn delete_migration_from_db_table<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        migration: &'life1 Box<dyn Migration<DB>>,
        connection: &'life2 mut <DB as Database>::Connection
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn fetch_applied_migration_from_db<'life0, 'life1, 'async_trait>(
        &'life0 self,
        connection: &'life1 mut <DB as Database>::Connection
    ) -> Pin<Box<dyn Future<Output = Result<Vec<AppliedMigrationSqlRow>, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn lock<'life0, 'life1, 'async_trait>(
        &'life0 self,
        connection: &'life1 mut <DB as Database>::Connection
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn unlock<'life0, 'life1, 'async_trait>(
        &'life0 self,
        connection: &'life1 mut <DB as Database>::Connection
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Trait which is implemented for database for performing different operations/action on database. Usually this trait is implemented for database to support database along with info trait

Required Methods

fn ensure_migration_table_exists<'life0, 'life1, 'async_trait>( &'life0 self, connection: &'life1 mut <DB as Database>::Connection ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Ensure migration table is created before running migrations. If not create one

fn drop_migration_table_if_exists<'life0, 'life1, 'async_trait>( &'life0 self, connection: &'life1 mut <DB as Database>::Connection ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Drop migration table if migration table exists

fn add_migration_to_db_table<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, migration: &'life1 Box<dyn Migration<DB>>, connection: &'life2 mut <DB as Database>::Connection ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Add migration to migration db table

fn delete_migration_from_db_table<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, migration: &'life1 Box<dyn Migration<DB>>, connection: &'life2 mut <DB as Database>::Connection ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Delete migration from migration db table

fn fetch_applied_migration_from_db<'life0, 'life1, 'async_trait>( &'life0 self, connection: &'life1 mut <DB as Database>::Connection ) -> Pin<Box<dyn Future<Output = Result<Vec<AppliedMigrationSqlRow>, Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

List all applied migrations from database as struct

fn lock<'life0, 'life1, 'async_trait>( &'life0 self, connection: &'life1 mut <DB as Database>::Connection ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Lock database while doing migrations so no two migrations run together

fn unlock<'life0, 'life1, 'async_trait>( &'life0 self, connection: &'life1 mut <DB as Database>::Connection ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Unlock locked database