Trait sqlx_migrator::migrator::Migrate

pub trait Migrate<DB>: Info<DB> + DatabaseOperation<DB> + Send + Syncwhere
    DB: Database,{
    // Provided methods
    fn list_applied_migrations<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<&'_ Box<dyn Migration<DB>>>, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn generate_migration_plan<'life0, 'async_trait>(
        &'life0 self,
        plan: Plan
    ) -> Pin<Box<dyn Future<Output = Result<Vec<&'_ Box<dyn Migration<DB>>>, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn apply_all<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn apply_migration<'life0, 'life1, 'async_trait>(
        &'life0 self,
        migration: &'life1 Box<dyn Migration<DB>>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn revert_all<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn revert_migration<'life0, 'life1, 'async_trait>(
        &'life0 self,
        migration: &'life1 Box<dyn Migration<DB>>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
}

Migrate trait which migrate a database according to requirements. This trait implements all methods which depends on DatabaseOperation trait. This trait is not required to implement any method since all have provided implementation and good to go for database agnostic case

Example

Create own custom Migrator which only supports postgres and uses own unique table name instead of default table name

use std::collections::HashSet;

use sqlx::{Pool, Postgres};
use sqlx_migrator::error::Error;
use sqlx_migrator::migration::{AppliedMigrationSqlRow, Migration};
use sqlx_migrator::migrator::{DatabaseOperation, Info, Migrate};

pub struct CustomMigrator {
    migrations: HashSet<Box<dyn Migration<Postgres>>>,
    pool: Pool<Postgres>,
}

impl Info<Postgres> for CustomMigrator {
    fn migrations(&self) -> &HashSet<Box<dyn Migration<Postgres>>> {
        &self.migrations
    }

    fn migrations_mut(&mut self) -> &mut HashSet<Box<dyn Migration<Postgres>>> {
        &mut self.migrations
    }

    fn pool(&self) -> &Pool<Postgres> {
        &self.pool
    }
}
#[must_use]
pub(crate) fn create_migrator_table_query() -> &'static str {
    "CREATE TABLE IF NOT EXISTS _custom_migrator_table_name (
       id INT PRIMARY KEY NOT NULL GENERATED ALWAYS AS IDENTITY,
       app TEXT NOT NULL,
       name TEXT NOT NULL,
       applied_time TIMESTAMPTZ NOT NULL DEFAULT now(),
       UNIQUE (app, name)
   )"
}

#[must_use]
pub(crate) fn drop_table_query() -> &'static str {
    "DROP TABLE IF EXISTS _custom_migrator_table_name"
}

#[must_use]
pub(crate) fn fetch_row_query() -> &'static str {
    "SELECT id, app, name, applied_time FROM _custom_migrator_table_name"
}

#[must_use]
pub(crate) fn add_migration_query() -> &'static str {
    "INSERT INTO _custom_migrator_table_name(app, name) VALUES ($1, $2)"
}

#[must_use]
pub(crate) fn delete_migration_query() -> &'static str {
    "DELETE FROM _custom_migrator_table_name WHERE app = $1 AND name = $2"
}

pub(crate) async fn lock_database(pool: &Pool<Postgres>) -> Result<(), Error> {
    let database = pool.connect_options().get_database().unwrap_or_default();
    let lock_id = i64::from(crc32fast::hash(database.as_bytes()));
    sqlx::query(
        "SELECT
pg_advisory_lock($1)",
    )
    .bind(lock_id)
    .execute(pool)
    .await?;
    Ok(())
}

pub(crate) async fn unlock_database(pool: &Pool<Postgres>) -> Result<(), Error> {
    let database = pool.connect_options().get_database().unwrap_or_default();
    let lock_id = i64::from(crc32fast::hash(database.as_bytes()));
    sqlx::query(
        "SELECT
pg_advisory_unlock($1)",
    )
    .bind(lock_id)
    .execute(pool)
    .await?;
    Ok(())
}

#[async_trait::async_trait]
impl DatabaseOperation<Postgres> for CustomMigrator {
    async fn ensure_migration_table_exists(&self) -> Result<(), Error> {
        sqlx::query(create_migrator_table_query())
            .execute(&self.pool)
            .await?;
        Ok(())
    }

    async fn drop_migration_table_if_exists(&self) -> Result<(), Error> {
        sqlx::query(drop_table_query()).execute(&self.pool).await?;
        Ok(())
    }

    async fn add_migration_to_db_table(
        &self,
        migration: &Box<dyn Migration<Postgres>>,
        connection: &mut <Postgres as sqlx::Database>::Connection,
    ) -> Result<(), Error> {
        sqlx::query(add_migration_query())
            .bind(migration.app())
            .bind(migration.name())
            .execute(connection)
            .await?;
        Ok(())
    }

    async fn delete_migration_from_db_table(
        &self,
        migration: &Box<dyn Migration<Postgres>>,
        connection: &mut <Postgres as sqlx::Database>::Connection,
    ) -> Result<(), Error> {
        sqlx::query(delete_migration_query())
            .bind(migration.app())
            .bind(migration.name())
            .execute(connection)
            .await?;
        Ok(())
    }

    async fn fetch_applied_migration_from_db(
        &self,
    ) -> Result<Vec<AppliedMigrationSqlRow>, Error> {
        let rows = sqlx::query_as(fetch_row_query())
            .fetch_all(&self.pool)
            .await?;
        Ok(rows)
    }

    async fn lock(&self) -> Result<(), Error> {
        lock_database(&self.pool).await
    }

    async fn unlock(&self) -> Result<(), Error> {
        unlock_database(&self.pool).await
    }
}

impl Migrate<Postgres> for CustomMigrator {}

Provided Methods

fn list_applied_migrations<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<&'_ Box<dyn Migration<DB>>>, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

List all applied migrations. Returns a vector of migration

fn generate_migration_plan<'life0, 'async_trait>( &'life0 self, plan: Plan ) -> Pin<Box<dyn Future<Output = Result<Vec<&'_ Box<dyn Migration<DB>>>, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Generate migration plan for according to plan type. Returns a vector of migration

fn apply_all<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Apply missing migration

Errors

If failed to apply migration

fn apply_migration<'life0, 'life1, 'async_trait>( &'life0 self, migration: &'life1 Box<dyn Migration<DB>> ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Apply given migration and add it to applied migration table

fn revert_all<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Revert all applied migration from database

Errors

If any migration or operation fails

fn revert_migration<'life0, 'life1, 'async_trait>( &'life0 self, migration: &'life1 Box<dyn Migration<DB>> ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Revert provided migration and remove migration from table