Trait sqlx_migrator::migrator::Migrator

pub trait Migrator: Send + Sync {
    type Database: Database;

    fn migrations(
        &self
    ) -> &HashSet<Box<dyn Migration<Database = Self::Database>>>;
    fn migrations_mut(
        &mut self
    ) -> &mut HashSet<Box<dyn Migration<Database = Self::Database>>>;
    fn pool(&self) -> &Pool<Self::Database>;
    fn ensure_migration_table_exists<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn add_migration_to_db_table<'t, 'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        migration_full_name: &'life1 str,
        transaction: &'life2 mut Transaction<'t, Self::Database>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        't: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn delete_migration_from_db_table<'t, 'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        migration_full_name: &'life1 str,
        transaction: &'life2 mut Transaction<'t, Self::Database>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        't: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn fetch_applied_migration_from_db<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<String>, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn add_migrations(
        &mut self,
        migrations: Vec<Box<dyn Migration<Database = Self::Database>>>
    ) { ... }
    fn add_migration(
        &mut self,
        migration: Box<dyn Migration<Database = Self::Database>>
    ) { ... }
    fn list_applied_migrations<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<&'_ Box<dyn Migration<Database = Self::Database>>>, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
    fn generate_migration_plan<'life0, 'async_trait>(
        &'life0 self,
        plan_type: PlanType
    ) -> Pin<Box<dyn Future<Output = Result<Vec<&'_ Box<dyn Migration<Database = Self::Database>>>, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
    fn apply_all<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
    fn apply_migration<'life0, 'life1, 'async_trait>(
        &'life0 self,
        migration: &'life1 Box<dyn Migration<Database = Self::Database>>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }
    fn revert_all<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
    fn revert_migration<'life0, 'life1, 'async_trait>(
        &'life0 self,
        migration: &'life1 Box<dyn Migration<Database = Self::Database>>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }

}

Migrator trait

Required Associated Types

type Database: Database

Database type

Required Methods

fn migrations(&self) -> &HashSet<Box<dyn Migration<Database = Self::Database>>>

Return migrations

fn migrations_mut(
&mut self
) -> &mut HashSet<Box<dyn Migration<Database = Self::Database>>>

Return mutable reference of migrations

fn pool(&self) -> &Pool<Self::Database>

Return pool of database

fn ensure_migration_table_exists<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Ensure migration table is created before running migrations. If not create one

fn add_migration_to_db_table<'t, 'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
migration_full_name: &'life1 str,
transaction: &'life2 mut Transaction<'t, Self::Database>
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
't: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Add migration to migration table

fn delete_migration_from_db_table<'t, 'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
migration_full_name: &'life1 str,
transaction: &'life2 mut Transaction<'t, Self::Database>
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
't: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Delete migration from table

fn fetch_applied_migration_from_db<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<String>, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

List all applied migrations from database in string format (full name of migration)

Provided Methods

fn add_migrations(
&mut self,
migrations: Vec<Box<dyn Migration<Database = Self::Database>>>
)

Add vector of migrations to Migrator object

fn add_migration(
&mut self,
migration: Box<dyn Migration<Database = Self::Database>>
)

Add single migration to migrator object

fn list_applied_migrations<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<&'_ Box<dyn Migration<Database = Self::Database>>>, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

List all applied migrations. Returns a vector of migration

fn generate_migration_plan<'life0, 'async_trait>(
&'life0 self,
plan_type: PlanType
) -> Pin<Box<dyn Future<Output = Result<Vec<&'_ Box<dyn Migration<Database = Self::Database>>>, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Generate migration plan for according to plan type. Returns a vector of migration

fn apply_all<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Apply missing migration

Errors

If failed to apply migration

fn apply_migration<'life0, 'life1, 'async_trait>(
&'life0 self,
migration: &'life1 Box<dyn Migration<Database = Self::Database>>
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Apply given migration and add it to applied migration table

fn revert_all<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Revert all applied migration from database

Errors

If any migration or operation fails

fn revert_migration<'life0, 'life1, 'async_trait>(
&'life0 self,
migration: &'life1 Box<dyn Migration<Database = Self::Database>>
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Revert provided migration and remove migration from table