Module sqlx_migrator::operation

Operation module

To create own operation implement trait for type

Example

use sqlx_migrator::error::Error;
use sqlx_migrator::operation::Operation;

struct ExampleOperation;
#[async_trait::async_trait]
impl Operation for ExampleOperation {
    type Database = sqlx_migrator::sqlx::Sqlite;

    async fn up(
        &self,
        connection: &mut <Self::Database as sqlx::Database>::Connection,
    ) -> Result<(), Error> {
        // Do some operations
        Ok(())
    }

    // By default operation is irreversible and cannot be reversed if you want to support
    // reverse of migration than add down function as well
    async fn down(
        &self,
        connection: &mut <Self::Database as sqlx::Database>::Connection,
    ) -> Result<(), Error> {
        // Do some operations
        Ok(())
    }
}

Traits

Operation

Trait for operation