Trait sqlx_migrator::operation::Operation

pub trait Operation<DB, State = ()>: Send + Sync
where
    DB: Database,
    State: Send + Sync,
{
    // Required method
    fn up<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        connection: &'life1 mut <DB as Database>::Connection,
        state: &'life2 State,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;

    // Provided methods
    fn down<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        connection: &'life1 mut <DB as Database>::Connection,
        state: &'life2 State,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn is_destructible(&self) -> bool { ... }
}

Trait for operation

Required Methods

fn up<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, connection: &'life1 mut <DB as Database>::Connection, state: &'life2 State, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Up command to be executed during migration apply

Provided Methods

fn down<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, connection: &'life1 mut <DB as Database>::Connection, state: &'life2 State, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Down command to be executed during migration rollback. If it is not implemented than operation is irreversible operation.

fn is_destructible(&self) -> bool

Whether up operation is destructible or not. If operation is destructible than user should answer before running migration through cli. By default up operation are false. Down operation are always destructible and cannot be changed