sqlx_migrator::sync

Trait OldMigrator

pub trait OldMigrator<DB>: Send + Sync
where
    DB: Database,
{
    // Required method
    fn applied_migrations<'life0, 'life1, 'async_trait>(
        &'life0 self,
        connection: &'life1 mut <DB as Database>::Connection,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Box<dyn Migration<DB>>>, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Trait which is implemented for syncing a migration from old migrator to new migrator

Required Methods

fn applied_migrations<'life0, 'life1, 'async_trait>( &'life0 self, connection: &'life1 mut <DB as Database>::Connection, ) -> Pin<Box<dyn Future<Output = Result<Vec<Box<dyn Migration<DB>>>, Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get a list of applied migration from old migrator which can be represented by sqlx_migrator migrator