sqlx_migrator

Trait Operation

pub trait Operation<DB>: Send + Sync
where
    DB: Database,
{
    // Required method
    fn up<'life0, 'life1, 'async_trait>(
        &'life0 self,
        connection: &'life1 mut <DB as Database>::Connection,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;

    // Provided methods
    fn down<'life0, 'life1, 'async_trait>(
        &'life0 self,
        connection: &'life1 mut <DB as Database>::Connection,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn is_destructible(&self) -> bool { ... }
}

Trait representing a database migration operation.

Operations are the building blocks of a migration, defining the specific changes to be made to the database schema or data. Each operation must implement the up method to apply the operation, and can optionally implement the down method to reverse the operation.

Required Methods

fn up<'life0, 'life1, 'async_trait>( &'life0 self, connection: &'life1 mut <DB as Database>::Connection, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

The up method applies the operation to the database.

This method is called when the migration is being applied to the database. Implement this method to define the changes you want to apply.

Provided Methods

fn down<'life0, 'life1, 'async_trait>( &'life0 self, connection: &'life1 mut <DB as Database>::Connection, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

The down method reverts the operation, if possible.

Down operations are optional and may not be implemented for all. If not implemented, the operation is considered irreversible. If you want to make an operation reversible, implement this method to define how to revert the changes made in the up method.

fn is_destructible(&self) -> bool

Indicates whether the up operation is destructible.

If the operation is destructible, the user will be prompted for confirmation before running the migration via the CLI, due to the potential for data loss or irreversible changes. By default, up operations are considered non-destructible. Note that down operations are always considered destructible and cannot be changed.