Crate sqlx_migrator

Library to create sqlx migration using rust code instead of sql.

Visit examples directory and readme file to learn of how to use sqlx migrator on sqlx project

Modules

clicli

Module for creating and running cli with help of migrator

error

Module for library error

migration

Module defining migration trait To create own implement migration trait for type

migrator

Migrator module

operation

Operation module To create own operation implement trait for type

Macros

any_migration(postgres or mysql or sqlite) and any

Macro for implementing the migration macro for the Any.

migration

Macro for implementing the Migration trait for the provided database.

mysql_migrationmysql

Macro for implementing the migration macro for the MySql.

postgres_migrationpostgres

Macro for implementing the migration macro for the Postgres.

sqlite_migrationsqlite

Macro for implementing the migration macro for the Sqlite.

vec_box

Macro for vector of Box