Crate sqlx_migrator

Library to create sqlx migration using rust code instead of sql. Visit examples and readme file to learn of how to use sqlx migrator on sqlx project

Re-exports

pub use sqlx;

Modules

clicli

Module for creating and running cli with help of migrator

error

Module for library error

migration

Module defining migration trait

migrator

Migrator module

operation

Operation module

Macros

any_migration(postgres or mysql or sqlite) and any

Macro for implementing the Migration trait for the Any.

any_operation(postgres or mysql or sqlite) and any

Macro for defining any SQL operations.

migration

Macro for implementing the Migration trait for the provided type.

mysql_migrationmysql

Macro for implementing the Migration trait for the MySql.

mysql_operationmysql

Macro for defining mysql SQL operations.

operation

Macro for defining SQL operations.

postgres_migrationpostgres

Macro for implementing the Migration trait for the Postgres.

postgres_operationpostgres

Macro for defining postgres SQL operations.

sqlite_migrationsqlite

Macro for implementing the Migration trait for the Sqlite.

sqlite_operationsqlite

Macro for defining sqlite SQL operations.

vec_box

Macro for vector of box