Crate sqlx_migrator

Library to create sqlx migration using rust code instead of sql

Re-exports

pub use sqlx;

Modules

clicli

Module for creating and running cli with help of migrator

error

Module for library error

migration

Module defining migration trait

migrator

migrator module

operation

Operation module

sqlitesqlite

sqlite module