Crate sqlx_migrator

Library to create sqlx migration using rust code instead of sql.

Check README.MD for more detailed information of how to use a crate and visit Operation, Migration and Migrator

Re-exports

pub use crate::cli::MigrationCommand;cli

pub use crate::error::Error;

pub use crate::migration::Migration;

pub use crate::migrator::Info;

pub use crate::migrator::Migrate;

pub use crate::migrator::Migrator;

pub use crate::migrator::Plan;

pub use crate::operation::Operation;

Modules

clicli

Module for creating and running cli with help of migrator

error

Module for library error

migration

Module for defining the Migration trait, which represents a database migration.

migrator

Migrator module

operation

Module for defining the Operation trait

Macros

any_migration(postgres or mysql or sqlite) and any

Macro for implementing the migration macro for the Any.

migration

Macro for implementing the Migration trait for the provided database.

mysql_migrationmysql

Macro for implementing the migration macro for the MySql.

postgres_migrationpostgres

Macro for implementing the migration macro for the Postgres.

sqlite_migrationsqlite

Macro for implementing the migration macro for the Sqlite.

vec_box

Macro for vector of Box