Macro sqlx_migrator::migration

macro_rules! migration {
    ($db:ty, $state:ty, $op:ty, $app_name:expr, $parents:expr, $operations:expr) => { ... };
    ($db:ty, $op:ty, $app_name:expr, $parents:expr, $operations:expr) => { ... };
}

Macro for implementing the Migration trait for the provided database.

This macro will use current file name as name for migration

This macro expects the following arguments:

$db:ty: the type of database

$state:ty: the type of state for migration (optional). If not present this will by default set as ()

$op:ty: The type for which the migration is being implemented

$app_name:expr: Name of app to be used for app variable

$parents:expr: List of parents migration.

$operations:expr: List of operations