Trait sqlx_migrator::migration::Migration

pub trait Migration<DB, State = ()>: Send + Sync {
    // Required methods
    fn app(&self) -> &str;
    fn name(&self) -> &str;
    fn parents(&self) -> Vec<Box<dyn Migration<DB, State>>>;
    fn operations(&self) -> Vec<Box<dyn Operation<DB, State>>>;

    // Provided methods
    fn replaces(&self) -> Vec<Box<dyn Migration<DB, State>>> { ... }
    fn run_before(&self) -> Vec<Box<dyn Migration<DB, State>>> { ... }
    fn is_atomic(&self) -> bool { ... }
}

Trait for migration

Required Methods

fn app(&self) -> &str

Migration app name. Can be name of folder or library where migration is located

fn name(&self) -> &str

Migration name. Can be file name without extension

fn parents(&self) -> Vec<Box<dyn Migration<DB, State>>>

Parents of migration (migrations that should be applied before this migration)

fn operations(&self) -> Vec<Box<dyn Operation<DB, State>>>

Operation performed for migration (create, drop, etc.)

Provided Methods

fn replaces(&self) -> Vec<Box<dyn Migration<DB, State>>>

Replace certain migrations. If any one of listed migration is applied than migration will not be applied else migration will apply/revert instead of applying/reverting those migration.

fn run_before(&self) -> Vec<Box<dyn Migration<DB, State>>>

Run before(for applying)/after(for reverting) certain migration. This can be helpful in condition where other library migration need to be applied after this migration or reverted before this migration

fn is_atomic(&self) -> bool

Whether migration is atomic or not. By default it is atomic so this function returns true