sqlx_migrator::migrator
Trait DatabaseOperation

pub trait DatabaseOperation<DB>
where
    DB: Database,
{
    // Required methods
    fn ensure_migration_table_exists<'life0, 'life1, 'async_trait>(
        &'life0 self,
        connection: &'life1 mut <DB as Database>::Connection,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn drop_migration_table_if_exists<'life0, 'life1, 'async_trait>(
        &'life0 self,
        connection: &'life1 mut <DB as Database>::Connection,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn add_migration_to_db_table<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        connection: &'life1 mut <DB as Database>::Connection,
        migration: &'life2 Box<dyn Migration<DB>>,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn delete_migration_from_db_table<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        connection: &'life1 mut <DB as Database>::Connection,
        migration: &'life2 Box<dyn Migration<DB>>,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn fetch_applied_migration_from_db<'life0, 'life1, 'async_trait>(
        &'life0 self,
        connection: &'life1 mut <DB as Database>::Connection,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<AppliedMigrationSqlRow>, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn lock<'life0, 'life1, 'async_trait>(
        &'life0 self,
        connection: &'life1 mut <DB as Database>::Connection,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn unlock<'life0, 'life1, 'async_trait>(
        &'life0 self,
        connection: &'life1 mut <DB as Database>::Connection,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

The DatabaseOperation trait defines a set of methods for performing operations related to migration management on the database.

This trait is typically implemented for a database to support migration operations, such as ensuring the migration table exists, adding or removing migrations from the table, and locking the database during migration processes.

Required Methods

fn ensure_migration_table_exists<'life0, 'life1, 'async_trait>( &'life0 self, connection: &'life1 mut <DB as Database>::Connection, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Ensure migration table is created before running migrations. If not create one

fn drop_migration_table_if_exists<'life0, 'life1, 'async_trait>( &'life0 self, connection: &'life1 mut <DB as Database>::Connection, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Drop migration table if migration table exists

fn add_migration_to_db_table<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, connection: &'life1 mut <DB as Database>::Connection, migration: &'life2 Box<dyn Migration<DB>>, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Adds a migration record to the migration table in the database.

fn delete_migration_from_db_table<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, connection: &'life1 mut <DB as Database>::Connection, migration: &'life2 Box<dyn Migration<DB>>, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Removes a migration record from the migration table in the database.

fn fetch_applied_migration_from_db<'life0, 'life1, 'async_trait>( &'life0 self, connection: &'life1 mut <DB as Database>::Connection, ) -> Pin<Box<dyn Future<Output = Result<Vec<AppliedMigrationSqlRow>, Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Fetches the list of applied migrations from the migration table in the database.

fn lock<'life0, 'life1, 'async_trait>( &'life0 self, connection: &'life1 mut <DB as Database>::Connection, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Lock database while doing migrations so no two migrations run together

fn unlock<'life0, 'life1, 'async_trait>( &'life0 self, connection: &'life1 mut <DB as Database>::Connection, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Unlock locked database