Trait sqlx_migrator::migrator::Info

pub trait Info<DB>: Send + Syncwhere
    DB: Database,{
    // Required methods
    fn migrations(&self) -> &HashSet<Box<dyn Migration<DB>>>;
    fn migrations_mut(&mut self) -> &mut HashSet<Box<dyn Migration<DB>>>;
    fn pool(&self) -> &Pool<DB>;

    // Provided methods
    fn add_migrations(&mut self, migrations: Vec<Box<dyn Migration<DB>>>) { ... }
    fn add_migration(&mut self, migration: Box<dyn Migration<DB>>) { ... }
}

Info trait which implements some of database agnostic methods to return data. Only required methods needs to be implemented if you want to create your own migrator struct. This trait implements methods which doesn’t depends on Database Operation trait methods

Required Methods

fn migrations(&self) -> &HashSet<Box<dyn Migration<DB>>>

Return migrations

fn migrations_mut(&mut self) -> &mut HashSet<Box<dyn Migration<DB>>>

Return mutable reference of migrations

fn pool(&self) -> &Pool<DB>

Return pool of database

Provided Methods

fn add_migrations(&mut self, migrations: Vec<Box<dyn Migration<DB>>>)

Add vector of migrations to Migrator object

fn add_migration(&mut self, migration: Box<dyn Migration<DB>>)

Add single migration to migrator object