Trait sqlx_migrator::migrator::Info

pub trait Info<DB, State> {
    // Required methods
    fn state(&self) -> &State;
    fn migrations(&self) -> &Vec<Box<dyn Migration<DB, State>>>;
    fn migrations_mut(&mut self) -> &mut Vec<Box<dyn Migration<DB, State>>>;

    // Provided methods
    fn add_migrations(&mut self, migrations: Vec<Box<dyn Migration<DB, State>>>) { ... }
    fn add_migration(&mut self, migration: Box<dyn Migration<DB, State>>) { ... }
}

Info trait which implements some of database agnostic methods to add migration or returns immutable or mutable migrations list

Required Methods

fn state(&self) -> &State

Return state used in migrator

fn migrations(&self) -> &Vec<Box<dyn Migration<DB, State>>>

Return migrations

fn migrations_mut(&mut self) -> &mut Vec<Box<dyn Migration<DB, State>>>

Return mutable reference of migrations

Provided Methods

fn add_migrations(&mut self, migrations: Vec<Box<dyn Migration<DB, State>>>)

Add vector of migrations to Migrator object

fn add_migration(&mut self, migration: Box<dyn Migration<DB, State>>)

Add single migration to migrator object