Trait sqlx_migrator::migrator::Migrate

pub trait Migrate<DB>: Info<DB> + DatabaseOperation<DB> + Send + Syncwhere
    DB: Database,{
    // Provided methods
    fn list_applied_migrations<'life0, 'life1, 'async_trait>(
        &'life0 self,
        connection: &'life1 mut <DB as Database>::Connection
    ) -> Pin<Box<dyn Future<Output = Result<Vec<&'_ Box<dyn Migration<DB>>>, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn generate_migration_plan<'life0, 'life1, 'async_trait>(
        &'life0 self,
        plan: Plan,
        connection: &'life1 mut <DB as Database>::Connection
    ) -> Pin<Box<dyn Future<Output = Result<Vec<&'_ Box<dyn Migration<DB>>>, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn apply_all<'life0, 'life1, 'async_trait>(
        &'life0 self,
        pool: &'life1 Pool<DB>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn apply_migration<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        migration: &'life1 Box<dyn Migration<DB>>,
        connection: &'life2 mut <DB as Database>::Connection
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn revert_all<'life0, 'life1, 'async_trait>(
        &'life0 self,
        pool: &'life1 Pool<DB>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn revert_migration<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        migration: &'life1 Box<dyn Migration<DB>>,
        connection: &'life2 mut <DB as Database>::Connection
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
}

Migrate trait which migrate a database according to requirements. This trait implements all methods which depends on DatabaseOperation trait and Info trait. This trait doesn’t requires to implement any method since all function have default implementation

Provided Methods

fn list_applied_migrations<'life0, 'life1, 'async_trait>( &'life0 self, connection: &'life1 mut <DB as Database>::Connection ) -> Pin<Box<dyn Future<Output = Result<Vec<&'_ Box<dyn Migration<DB>>>, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

List all applied migrations. Returns a vector of migration

fn generate_migration_plan<'life0, 'life1, 'async_trait>( &'life0 self, plan: Plan, connection: &'life1 mut <DB as Database>::Connection ) -> Pin<Box<dyn Future<Output = Result<Vec<&'_ Box<dyn Migration<DB>>>, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Generate migration plan for according to plan type. Returns a vector of migration

fn apply_all<'life0, 'life1, 'async_trait>( &'life0 self, pool: &'life1 Pool<DB> ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Apply all migrations which are not applied till now

Errors

If failed to apply migration

fn apply_migration<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, migration: &'life1 Box<dyn Migration<DB>>, connection: &'life2 mut <DB as Database>::Connection ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Apply given migration and add it to applied migration table

fn revert_all<'life0, 'life1, 'async_trait>( &'life0 self, pool: &'life1 Pool<DB> ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Revert all applied migration from database

Errors

If any migration or operation fails

fn revert_migration<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, migration: &'life1 Box<dyn Migration<DB>>, connection: &'life2 mut <DB as Database>::Connection ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Revert provided migration and remove migration from table