Trait sqlx_migrator::migrator::MigratorTrait

pub trait MigratorTrait<DB>: Send + Syncwhere
    DB: Database,{

    fn migrations(&self) -> &HashSet<Box<dyn MigrationTrait<DB>>>;
    fn migrations_mut(&mut self) -> &mut HashSet<Box<dyn MigrationTrait<DB>>>;
    fn pool(&self) -> &Pool<DB>;
    fn ensure_migration_table_exists<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn add_migration_to_db_table<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        migration: &'life1 Box<dyn MigrationTrait<DB>>,
        connection: &'life2 mut <DB as Database>::Connection
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn delete_migration_from_db_table<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        migration: &'life1 Box<dyn MigrationTrait<DB>>,
        connection: &'life2 mut <DB as Database>::Connection
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn fetch_applied_migration_from_db<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<AppliedMigrationSqlRow>, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn add_migrations(&mut self, migrations: Vec<Box<dyn MigrationTrait<DB>>>) { ... }
    fn add_migration(&mut self, migration: Box<dyn MigrationTrait<DB>>) { ... }
    fn list_applied_migrations<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<&'_ Box<dyn MigrationTrait<DB>>>, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
    fn generate_migration_plan<'life0, 'async_trait>(
        &'life0 self,
        plan: Plan
    ) -> Pin<Box<dyn Future<Output = Result<Vec<&'_ Box<dyn MigrationTrait<DB>>>, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
    fn apply_all<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
    fn apply_migration<'life0, 'life1, 'async_trait>(
        &'life0 self,
        migration: &'life1 Box<dyn MigrationTrait<DB>>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }
    fn revert_all<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
    fn revert_migration<'life0, 'life1, 'async_trait>(
        &'life0 self,
        migration: &'life1 Box<dyn MigrationTrait<DB>>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }

}

Migrator trait

Required Methods

fn migrations(&self) -> &HashSet<Box<dyn MigrationTrait<DB>>>

Return migrations

fn migrations_mut(&mut self) -> &mut HashSet<Box<dyn MigrationTrait<DB>>>

Return mutable reference of migrations

fn pool(&self) -> &Pool<DB>

Return pool of database

fn ensure_migration_table_exists<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Ensure migration table is created before running migrations. If not create one

fn add_migration_to_db_table<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
migration: &'life1 Box<dyn MigrationTrait<DB>>,
connection: &'life2 mut <DB as Database>::Connection
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Add migration to migration table

fn delete_migration_from_db_table<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
migration: &'life1 Box<dyn MigrationTrait<DB>>,
connection: &'life2 mut <DB as Database>::Connection
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Delete migration from table

fn fetch_applied_migration_from_db<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<AppliedMigrationSqlRow>, Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

List all applied migrations from database in string format (full name of migration)

Provided Methods

fn add_migrations(&mut self, migrations: Vec<Box<dyn MigrationTrait<DB>>>)

Add vector of migrations to Migrator object

fn add_migration(&mut self, migration: Box<dyn MigrationTrait<DB>>)

Add single migration to migrator object

fn list_applied_migrations<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<&'_ Box<dyn MigrationTrait<DB>>>, Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

List all applied migrations. Returns a vector of migration

fn generate_migration_plan<'life0, 'async_trait>(
&'life0 self,
plan: Plan
) -> Pin<Box<dyn Future<Output = Result<Vec<&'_ Box<dyn MigrationTrait<DB>>>, Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Generate migration plan for according to plan type. Returns a vector of migration

fn apply_all<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Apply missing migration

Errors

If failed to apply migration

fn apply_migration<'life0, 'life1, 'async_trait>(
&'life0 self,
migration: &'life1 Box<dyn MigrationTrait<DB>>
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Apply given migration and add it to applied migration table

fn revert_all<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Revert all applied migration from database

Errors

If any migration or operation fails

fn revert_migration<'life0, 'life1, 'async_trait>(
&'life0 self,
migration: &'life1 Box<dyn MigrationTrait<DB>>
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Revert provided migration and remove migration from table